# cluster_manager/apiserver/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this service is to provide an application\nthat is using plain go code to define an API",
    "title": "HTTP API.",
    "contact": {
      "name": "Hubery",
      "email": "fuguangrong@bsgchina.com"
    },
    "version": "1.0"
  },
  "host": "localhost",
  "basePath": "/v1.0",
  "paths": {
...
    "/manager/tasks": {
      "get": {
        "description": "List Tasks\nThis will returns a list of tasks by options",
        "tags": [
          "tasks"
        ],
        "summary": "查询任务列表",
        "operationId": "listTasks",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "RelateID",
            "name": "related_id",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Action",
            "name": "action",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Status",
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/listTasksResponseWrapper"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Task": {
      "type": "object",
      "properties": {
        "action": {
          "type": "string",
          "x-go-name": "Action"
        },
        "created_at": {
          "$ref": "#/definitions/Time"
        },
        "created_user": {
          "type": "string",
          "x-go-name": "User"
        },
        "error": {
          "type": "string",
          "x-go-name": "Error"
        },
        "finished_at": {
          "$ref": "#/definitions/Time"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "relate_id": {
          "type": "string",
          "x-go-name": "RelateID"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/upmio/dbscale-kube/cluster_manager/apiserver/api"
    },
...
    "TasksResponse": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Task"
      },
      "x-go-package": "github.com/upmio/dbscale-kube/cluster_manager/apiserver/api"
    },
    "Time": {
      "type": "string",
      "format": "date-time",
      "x-go-package": "github.com/upmio/dbscale-kube/cluster_manager/apiserver/api"
    },
...
  },
  "responses": {
    "ErrorResponse": {
      "description": "ErrorResponse\nerror common response",
      "headers": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "错误代码"
        },
        "msg": {
          "type": "string",
          "description": "错误信息"
        }
      }
    },
...
    "listTasksResponseWrapper": {
      "description": "list tasks info",
      "schema": {
        "$ref": "#/definitions/TasksResponse"
      }
    }
  }
}